Added invalid list exceptions to jobs

+++ src/Models/List/Jobs/ListDeleteJob.php
@@ -4,6 +4,8 @@ class ListDeleteJob extends AbstractJob
 	public function execute(UserEntity $owner)
 	{
 		$listEntity = ListService::getByUrlName($owner, $this->arguments['list-id']);
+		if (empty($listEntity))
+			throw new InvalidListException($this->arguments['list-id']);
 
 		ListService::delete($listEntity);
 	}
